:bug: Fixes reader bug for local_to_global files
In https://github.com/schism-dev/schism/commit/16c366e element side
mapping was added to the end of local_to_global to be used with ViSit
plugin. As the mergegr3 script was using a bottom-to-top slicing
strategy, it breaks the reading of element and nodes table. This bug is
fixed by checking the total length of the file and comparing with the
computed length.